# perf fixes copy from dotutils and MSBuildStructuredLog