Fixes #10674 #10675

### Context
https://github.com/dotutils/streamutils/pull/2 makes more performant custom implementations of streams which can be copied here

https://github.com/KirillOsenkov/MSBuildStructuredLog/pull/821 indroduced changes to BuildEventArgsReader which should be the same in MSBuildStructuredLog, this PR copies these changes.

### Changes Made
3 commits
1. override custom stream methods
2. adopt changes to BuildEventArgsReader
3. make non-params versions of resource methods

### Testing
perf improvement validation setup:
1. get validation .binlog from msbuild's own build on this branch `.\build /bl /p:Configuration=Release, 
4. also do this for a baseline from main [94941d9](https://github.com/dotnet/msbuild/commit/94941d9cb26bb86045452b4a174a357b65a30c99)
5. measurement of replay speed in powershell script: run with each version bootstrapped replay `versionX\artifacts\bin\bootstrap\core\dotnet.exe build ThisPRVersion\msbuild.binlog`
6. 100x measurements for both versions alternating their runs
7. mean, stdev, t-test

Results on Devbox machine:

Command 1: `.\msbuild\artifacts\bin\bootstrap\core\dotnet.exe build .\msbuild\msbuild.binlog`
Average runtime: 1790.30201 ms
Standard deviation: 247.701794004614 ms

Command 2: `.\msbuildbase\artifacts\bin\bootstrap\core\dotnet.exe build .\msbuild\msbuild.binlog`
Average runtime: 2202.75113 ms
Standard deviation: 146.88933492219 ms

T-statistic: -14.3221294766752
degrees of freedom 100-1
pvalue <0.0001, we can reject the hypothesis that the distributions have the same mean, the alternate hypothesis is that the change is faster.

_We got a ~23% speedup for this scenario._

(even better results on OrchardCore, but less rigorously obtained)
### Notes